+++ addons/sevenrings/Eastern/CfgVehicles_Eastern_Turrets.hpp
@@ -6,60 +6,6 @@ class TACU_Seven_Rings_V_O_MK6: TACU_Assets_MK6_Base {
     typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
 };
 
-class TACU_Seven_Rings_V_O_AGS30: TACU_Assets_AGS_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Seven_Rings_Eastern";
-    side = 0;
-    crew = "TACU_Seven_Rings_Eastern_U_O_Rifleman_01";
-    typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
-};
-
-class TACU_Seven_Rings_V_O_D30: TACU_Assets_D30_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Seven_Rings_Eastern";
-    side = 0;
-    crew = "TACU_Seven_Rings_Eastern_U_O_Rifleman_01";
-    typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
-};
-
-class TACU_Seven_Rings_V_O_DSHKM: TACU_Assets_DSHKM_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Seven_Rings_Eastern";
-    side = 0;
-    crew = "TACU_Seven_Rings_Eastern_U_O_Rifleman_01";
-    typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
-};
-class TACU_Seven_Rings_V_O_DSHKM_Mini: TACU_Assets_DSHKM_Mini_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Seven_Rings_Eastern";
-    side = 0;
-    crew = "TACU_Seven_Rings_Eastern_U_O_Rifleman_01";
-    typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
-};
-
-class TACU_Seven_Rings_V_O_Igla: TACU_Assets_IGLA_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Seven_Rings_Eastern";
-    side = 0;
-    crew = "TACU_Seven_Rings_Eastern_U_O_Rifleman_01";
-    typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
-};
-
-class TACU_Seven_Rings_V_O_KORD: TACU_Assets_KORD_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Seven_Rings_Eastern";
-    side = 0;
-    crew = "TACU_Seven_Rings_Eastern_U_O_Rifleman_01";
-    typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
-};
-class TACU_Seven_Rings_V_O_KORD_Mini: TACU_Assets_KORD_Mini_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Seven_Rings_Eastern";
-    side = 0;
-    crew = "TACU_Seven_Rings_Eastern_U_O_Rifleman_01";
-    typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
-};
-
 class TACU_Seven_Rings_V_O_ZU23: TACU_Assets_ZU23_base {
     MACRO_FACTION_VEHICLE_DEFAULTS;
     faction = "TACU_Seven_Rings_Eastern";
