+++ addons/sevenrings/Western/CfgVehicles_Western_Turrets.hpp
@@ -6,31 +6,7 @@ class TACU_Seven_Rings_V_I_HMG: TACU_Assets_HMG_50 {
     typicalCargo[] = {"TACU_Seven_Rings_Western_U_I_Rifleman_01"};
 };
 
-class TACU_Seven_Rings_V_I_Metis: TACU_Assets_Metis_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Seven_Rings_Western";
-    side = 2;
-    crew = "TACU_Seven_Rings_Western_U_I_Rifleman_01";
-    typicalCargo[] = {"TACU_Seven_Rings_Western_U_I_Rifleman_01"};
-};
-
-class TACU_Seven_Rings_V_I_MK19: TACU_Assets_MK19_base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Seven_Rings_Western";
-    side = 2;
-    crew = "TACU_Seven_Rings_Western_U_I_Rifleman_01";
-    typicalCargo[] = {"TACU_Seven_Rings_Western_U_I_Rifleman_01"};
-};
-
 class TACU_Seven_Rings_V_I_MK6: TACU_Assets_MK6_Base {
-    MACRO_FACTION_VEHICLE_DEFAULTS;
-    faction = "TACU_Seven_Rings_Western";
-    side = 2;
-    crew = "TACU_Seven_Rings_Western_U_I_Rifleman_01";
-    typicalCargo[] = {"TACU_Seven_Rings_Western_U_I_Rifleman_01"};
-};
-
-class TACU_Seven_Rings_V_I_TOW: TACU_Assets_M220_TOW_base {
     MACRO_FACTION_VEHICLE_DEFAULTS;
     faction = "TACU_Seven_Rings_Western";
     side = 2;
