+++ addons/sevenrings/subconfig_cup/CfgVehicles.hpp
@@ -1,0 +1,83 @@
+class CfgVehicles {
+    #include "\x\tacu\addons\assets\script_classes_vehicles_cup.hpp"
+
+    // Eastern
+    class TACU_Seven_Rings_V_O_AGS30: TACU_Assets_AGS_base {
+        MACRO_FACTION_VEHICLE_DEFAULTS;
+        faction = "TACU_Seven_Rings_Eastern";
+        side = 0;
+        crew = "TACU_Seven_Rings_Eastern_U_O_Rifleman_01";
+        typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
+    };
+
+    class TACU_Seven_Rings_V_O_D30: TACU_Assets_D30_base {
+        MACRO_FACTION_VEHICLE_DEFAULTS;
+        faction = "TACU_Seven_Rings_Eastern";
+        side = 0;
+        crew = "TACU_Seven_Rings_Eastern_U_O_Rifleman_01";
+        typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
+    };
+
+    class TACU_Seven_Rings_V_O_DSHKM: TACU_Assets_DSHKM_base {
+        MACRO_FACTION_VEHICLE_DEFAULTS;
+        faction = "TACU_Seven_Rings_Eastern";
+        side = 0;
+        crew = "TACU_Seven_Rings_Eastern_U_O_Rifleman_01";
+        typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
+    };
+    class TACU_Seven_Rings_V_O_DSHKM_Mini: TACU_Assets_DSHKM_Mini_base {
+        MACRO_FACTION_VEHICLE_DEFAULTS;
+        faction = "TACU_Seven_Rings_Eastern";
+        side = 0;
+        crew = "TACU_Seven_Rings_Eastern_U_O_Rifleman_01";
+        typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
+    };
+
+    class TACU_Seven_Rings_V_O_Igla: TACU_Assets_IGLA_base {
+        MACRO_FACTION_VEHICLE_DEFAULTS;
+        faction = "TACU_Seven_Rings_Eastern";
+        side = 0;
+        crew = "TACU_Seven_Rings_Eastern_U_O_Rifleman_01";
+        typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
+    };
+
+    class TACU_Seven_Rings_V_O_KORD: TACU_Assets_KORD_base {
+        MACRO_FACTION_VEHICLE_DEFAULTS;
+        faction = "TACU_Seven_Rings_Eastern";
+        side = 0;
+        crew = "TACU_Seven_Rings_Eastern_U_O_Rifleman_01";
+        typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
+    };
+    class TACU_Seven_Rings_V_O_KORD_Mini: TACU_Assets_KORD_Mini_base {
+        MACRO_FACTION_VEHICLE_DEFAULTS;
+        faction = "TACU_Seven_Rings_Eastern";
+        side = 0;
+        crew = "TACU_Seven_Rings_Eastern_U_O_Rifleman_01";
+        typicalCargo[] = {"TACU_Seven_Rings_Eastern_U_O_Rifleman_01"};
+    };
+
+    // Western
+    class TACU_Seven_Rings_V_I_Metis: TACU_Assets_Metis_base {
+        MACRO_FACTION_VEHICLE_DEFAULTS;
+        faction = "TACU_Seven_Rings_Western";
+        side = 2;
+        crew = "TACU_Seven_Rings_Western_U_I_Rifleman_01";
+        typicalCargo[] = {"TACU_Seven_Rings_Western_U_I_Rifleman_01"};
+    };
+
+    class TACU_Seven_Rings_V_I_MK19: TACU_Assets_MK19_base {
+        MACRO_FACTION_VEHICLE_DEFAULTS;
+        faction = "TACU_Seven_Rings_Western";
+        side = 2;
+        crew = "TACU_Seven_Rings_Western_U_I_Rifleman_01";
+        typicalCargo[] = {"TACU_Seven_Rings_Western_U_I_Rifleman_01"};
+    };
+
+    class TACU_Seven_Rings_V_I_TOW: TACU_Assets_M220_TOW_base {
+        MACRO_FACTION_VEHICLE_DEFAULTS;
+        faction = "TACU_Seven_Rings_Western";
+        side = 2;
+        crew = "TACU_Seven_Rings_Western_U_I_Rifleman_01";
+        typicalCargo[] = {"TACU_Seven_Rings_Western_U_I_Rifleman_01"};
+    };
+};
